Split Create invite commands into two different commands and other review comment fixes
