+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Invite/CreateEventManagerInviteCommand.cs
@@ -3,7 +3,7 @@ using MediatR;
 
 namespace AllReady.Areas.Admin.Features.Invite
 {
-    public class CreateInviteCommand : IAsyncRequest
+    public class CreateEventManagerInviteCommand : IAsyncRequest
     {
         public InviteViewModel Invite { get; set; }
         public string UserId { get; set; }
