+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Views/Invite/Send.cshtml
@@ -1,23 +1,13 @@
 ﻿@model AllReady.Areas.Admin.ViewModels.Invite.InviteViewModel
 
 @{
-    string formAction;
-    if (Model != null && Model.InviteType == AllReady.Areas.Admin.ViewModels.Invite.InviteType.EventManagerInvite)
-    {
-        ViewData["Title"] = "Send Event Manager Invite";
-        formAction = "SendEventManagerInvite";
-    }
-    else
-    {
-        ViewData["Title"] = "Send Campaign Manager Invite";
-        formAction = "SendCampaignManagerInvite";
-    }
+    ViewBag.Title = Model.Title;
 }
 
 <h2>@ViewData["Title"]</h2>
 <div class="row">
     <div class="col-12">
-        <form class="form-horizontal" asp-controller="Invite" asp-area="Admin" asp-action="@formAction" method="post" role="form">
+        <form class="form-horizontal" asp-controller="Invite" asp-area="Admin" asp-action="@Model.FormAction" method="post" role="form">
             <div asp-validation-summary="All" class="text-danger"></div>
             <div class="form-group">
                 <label asp-for="InviteeEmailAddress" class="col-md-2 control-label"></label>
@@ -40,8 +30,6 @@
                     <button type="submit" class="btn btn-default">Send</button>
                 </div>
             </div>
-            
-            <input type="hidden" asp-for="InviteType" />
         </form>
     </div>
 </div>
